build-image/requirements.txt
docker
gitpython
./cirpa/cirpa-utils
